Use ESM syntax, and update deps to support it
This switches from the commonjs require() syntax to the newer ESM
import syntax.

Jest did not work with the ESM syntax, so I updated the Jest config to
use esbuild to transpile the ESM syntax to commonjs, so Jest can still
run the tests.

We do this in the main CO2.js package, to it seemed an ok precedent to
follow.

This also updates better-sqlite, to use a new version which seems to
offer pre-builds, as of #1002 on the better sqlite repo.
https://github.com/WiseLibs/better-sqlite3/pull/1002